requirements.txt
Django==6.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
pillow==10.3.0
django-crispy-forms==2.0
crispy-bootstrap5==2024.2
